// src/services/articleShape/articleShape.types.ts
// -------------------------------------------------------------------------------------------

export interface ArticleShapeInterface {
	articleShapeId?: number;
	articleShape: string;
	companyId?: number;
}

// -------------------------------------------------------------------------------------------